+++ next.config.ts
@@ -11,16 +11,8 @@ const nextConfig: NextConfig = {
         protocol: "https",
         hostname: "cdn.auth0.com",
       },
-      // for displaying image urls inputed by users, temporary fix
-      {
-        protocol: "https",
-        hostname: "**", // Allows all hostnames with HTTPS
-        port: "",
-        pathname: "/**", // Allows all paths
-      },
     ],
   },
 };
 
 export default nextConfig;
-
