+++ src/app/(dashboard)/dashboard/career/resumes/page.tsx
@@ -37,7 +37,7 @@ export default function ResumesPage() {
 
   if (isLoading)
     return (
-      <div className="flex flex-wrap gap-6">
+      <div className="flex gap-6">
         <Skeleton.Node style={{ width: "288px", height: "392px" }} active />
         <Skeleton.Node style={{ width: "288px", height: "392px" }} active />
         <Skeleton.Node style={{ width: "288px", height: "392px" }} active />
@@ -68,8 +68,7 @@ export default function ResumesPage() {
 
   const handleDeleteResume = async (resumeId: string) => {
     await del(`/api/resumes/${resumeId}`);
-    mutate();
-    // data?.splice(data.indexOf(resumeData), 1);
+    data?.splice(data.indexOf(resumeData), 1);
   };
 
   return (
@@ -102,7 +101,7 @@ export default function ResumesPage() {
                 <Card
                   bordered={false}
                   size="small"
-                  key={resume.id}
+                  key={i}
                   className="w-72"
                   actions={[
                     <EditOutlined
@@ -110,7 +109,7 @@ export default function ResumesPage() {
                       onClick={() => handleResumeEditClick(resume)}
                     />,
                     <Popconfirm
-                      key={resume.id}
+                      key={i}
                       title="Delete the resume"
                       description="Are you sure you want to delete this resume?"
                       onConfirm={() => handleDeleteResume(resume.id!)}
