+++ src/components/dashboard/career/cv-builder/ResumeBuilder.tsx
@@ -149,28 +149,27 @@ const ResumeBuilder: React.FC<Props> = ({ setShowCvBuilder }) => {
       {contextHolder}
       <div className="flex flex-col xl:flex-row gap-5 justify-center">
         <div className="flex-grow xl:max-w-[460px] 2xl:max-w-[500px]">
-          <div className="flex flex-wrap items-center gap-5 sm:justify-between">
+          <div className="flex items-center justify-between">
             <button
               onMouseDown={() => {
                 setShowCvBuilder(false);
                 mutate(`/api/resumes`);
               }}
-              className="bg-white w-full max-w-[200px] border border-[#CBCBCB] min-h-12 rounded flex items-center justify-center gap-[7px] cursor-pointer"
+              className="my-4 bg-white w-full max-w-[200px] border border-[#CBCBCB] min-h-12 rounded flex items-center justify-center gap-[7px] cursor-pointer"
             >
               <LeftOutlined size={24} color="#010309" />
               <span className="text-[#010309] font-medium text-[13px] font-poppins">
                 Back to Templates
               </span>
             </button>
-            <button className="bg-[#010B18] w-full max-w-[200px] border border-[#010B18] min-h-12 rounded flex items-center justify-center gap-[7px] cursor-pointer">
+            <button className="my-4 bg-[#010B18] w-full max-w-[200px] border border-[#010B18] min-h-12 rounded flex items-center justify-center gap-[7px] cursor-pointer">
               <span className="text-white font-medium text-[13px] font-poppins">
                 Download
               </span>
               <FileDown size={18} color="#ffffff" />
             </button>
           </div>
-
-          <div className=" bg-white p-1 sm:p-5 sticky top-0 max-h-[calc(100vh-4rem)] overflow-y-auto mt-5">
+          <div className=" bg-white p-5 sticky top-0 max-h-[calc(100vh-4rem)] overflow-y-auto">
             <Collapse
               ghost
               items={items}
