+++ src/components/dashboard/career/templates/ModernTemplate.tsx
@@ -1,8 +1,5 @@
 import { ResumeType } from "@/app/types/career";
 import { Divider, Space, Typography } from "antd";
-import dayjs from "dayjs";
-import { Mail, MapPin, Phone, Link as LucideLink } from "lucide-react";
-import Link from "next/link";
 import React from "react";
 const { Title, Text } = Typography;
 
@@ -23,87 +20,47 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
   } = data;
 
   return (
-    <div className="bg-white text-gray-800 min-h-[1123px] p-5">
+    <div className="bg-white text-gray-800 min-h-[1123px]">
       {/* BASICS SECTION */}
-      <header className="">
-        <section className="flex flex-wrap gap-x-5 gap-y-2.5 items-end">
-          <h2 className="!font-dm_san text-3xl font-semibold">
-            {basics?.name ?? ""}
-          </h2>
-          <span className="!font-dm_sans text-base">
-            {basics?.label ?? <Text type="secondary"></Text>}
-          </span>
-        </section>
-        <section className="mt-2 grid grid-cols-1 sm:grid-cols-2 gap-2.5 !font-dm_sans">
-          {basics?.email && (
-            <span className="text-sm flex items-center gap-0.5">
-              {" "}
-              <Mail
-                stroke="#ffffff"
-                size={14}
-                fill="#000000"
-                className="mt-0.5"
-              />{" "}
-              {basics.email}
-            </span>
-          )}
-          {basics?.phone && (
-            <span className="text-sm flex items-center gap-1">
-              {" "}
-              <Phone
-                stroke="#ffffff"
-                size={14}
-                fill="#000000"
-                // className="mt-0.5"
-              />{" "}
-              {basics.phone}
-            </span>
-          )}
+      <header className="mb-10">
+        <Title level={1} style={{ marginBottom: 0 }}>
+          {basics?.name ?? ""}
+        </Title>
+        {basics?.label ?? <Text type="secondary"></Text>}
+        <br />
+        <Space direction="horizontal" className="mt-2">
+          {basics?.email && <Text>{basics.email}</Text>}
+          {basics?.phone && <Text>{basics.phone}</Text>}
           {basics?.url && (
-            <span className="flex sm:items-center gap-0.5">
-              <LucideLink size={14} color="#000000" />
-              <Link
-                href={basics.url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-blue-600"
-              >
-                {basics.url}
-              </Link>
-            </span>
+            <a
+              href={basics.url}
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-blue-600"
+            >
+              {basics.url}
+            </a>
           )}
-        </section>
-
-        {basics?.location && (
-          <div className="mt-2.5 flex items-center gap-0.5">
-            <MapPin stroke="#ffffff" size={18} fill="#000000" />
+        </Space>
+        <div className="mt-2">
+          {basics?.location && (
             <Text>
               {basics.location.address && `${basics.location.address}, `}
               {basics.location.city && `${basics.location.city}, `}
               {basics.location.region && `${basics.location.region}, `}
               {basics.location.countryCode && basics.location.countryCode}
             </Text>
+          )}
+        </div>
+        {basics?.summary && (
+          <div className="mt-4">
+            <Text>{basics.summary}</Text>
           </div>
         )}
-        {basics?.summary && (
-          <section className="mt-8 flex flex-col gap-2.5">
-            <div>
-              <h3 className="text-2xl text-black font-medium font-dm_sans">
-                Summary
-              </h3>
-              <div className="h-0.5 w-full border-b-2 border-b-black" />
-            </div>
-            <Text className="!font-dm_sans">{basics.summary}</Text>
-          </section>
-        )}
         {basics?.profiles && basics.profiles.length > 0 && (
-          <div className="mt-8 flex flex-col gap-2.5">
-            <div>
-              <h3 className="text-2xl text-black font-medium font-dm_sans">
-                Profiles
-              </h3>
-              <div className="h-0.5 w-full border-b-2 border-b-black" />
-            </div>
+          <div className="mt-4">
+            <Title level={4}>Profiles</Title>
+            <Divider />
             <ul className="list-none pl-0">
               {basics.profiles.map((profile, i) => (
                 <li key={i} className="mb-2">
@@ -114,7 +71,7 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
                     rel="noopener noreferrer"
                     className="text-blue-600"
                   >
-                    {profile?.url}
+                    {profile?.username}
                   </a>
                 </li>
               ))}
@@ -125,51 +82,32 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* WORK EXPERIENCE SECTION */}
       {work && work.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Work Experience
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Work Experience</Title>
+          <Divider />
           {work.map((job, i) => (
-            <div key={i} className="mb-6 mt-2.5">
-              <Text className="!font-dm_sans" strong>
-                {job?.position || "Position not specified"}
-              </Text>{" "}
-              <Text className="!font-dm_sans">
-                at {job?.name || "Company not specified"}
-              </Text>
+            <div key={i} className="mb-6">
+              <Text strong>{job.position}</Text> <Text>at {job.name}</Text>
               <br />
-              <Text className="!font-dm_sans" type="secondary">
-                {job?.startDate
-                  ? dayjs(job.startDate).format("MMM YYYY")
-                  : "Start date not specified"}{" "}
-                -{" "}
-                {job?.endDate
-                  ? dayjs(job.endDate).format("MMM YYYY")
-                  : "Present"}
+              <Text type="secondary">
+                {String(job.startDate)} - {String(job.endDate) || "Present"}
               </Text>
-              {job?.url && (
+              {job.url && (
                 <>
-                  <span className="flex sm:items-center gap-1">
-                    <LucideLink size={14} color="#000000" />
-                    <Link
-                      href={job.url}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="text-blue-600"
-                    >
-                      {job.url}
-                    </Link>
-                  </span>
+                  <br />
+                  <a
+                    href={job.url}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-blue-600"
+                  >
+                    {job.url}
+                  </a>
                 </>
               )}
-              {job?.summary && (
-                <p className="mt-2 font-dm_sans">{job.summary}</p>
-              )}
-              {job?.highlights && job.highlights.length > 0 && (
-                <ul className="list-disc list-inside mt-2 font-dm_sans">
+              {job.summary && <p className="mt-2">{job.summary}</p>}
+              {job.highlights && job.highlights.length > 0 && (
+                <ul className="list-disc list-inside mt-2">
                   {job.highlights.map((hl, idx) => (
                     <li key={idx}>{hl}</li>
                   ))}
@@ -182,14 +120,9 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* VOLUNTEER SECTION */}
       {volunteer && volunteer.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Volunteer Experience
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
-
+        <section className="mt-10">
+          <Title level={3}>Volunteer Experience</Title>
+          <Divider />
           {volunteer.map((vol, i) => (
             <div key={i} className="mb-6">
               <Text strong>{vol.position}</Text>{" "}
@@ -226,40 +159,34 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* EDUCATION SECTION */}
       {education && education.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Education
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Education</Title>
+          <Divider />
           {education.map((edu, i) => (
-            <div key={i} className="mb-6 mt-2.5">
+            <div key={i} className="mb-6">
               <Text strong>
-                {edu?.studyType} in {edu?.area}
+                {edu.studyType} in {edu.area}
               </Text>{" "}
-              at <Text strong>{edu?.institution}</Text>
+              at <Text strong>{edu.institution}</Text>
               <br />
               <Text type="secondary">
-                {dayjs(edu?.startDate).format("MMM YYYY")} -{" "}
-                {dayjs(edu?.endDate).format("MMM YYYY") || "Present"}
+                {String(edu.startDate)} - {String(edu.endDate) || "Present"}
               </Text>
-              {edu?.url && (
-                <div className="flex items-center gap-0.5">
-                  <LucideLink size={14} color="#000000" />
+              {edu.url && (
+                <>
                   <br />
-                  <Link
-                    href={edu?.url}
+                  <a
+                    href={edu.url}
                     target="_blank"
                     rel="noopener noreferrer"
                     className="text-blue-600"
                   >
-                    {edu?.url}
-                  </Link>
-                </div>
+                    {edu.url}
+                  </a>
+                </>
               )}
-              {edu?.score && <p className="mt-2">Score: {edu.score}</p>}
-              {edu?.courses && edu.courses.length > 0 && (
+              {edu.score && <p className="mt-2">Score: {edu.score}</p>}
+              {edu.courses && edu.courses.length > 0 && (
                 <div className="mt-2">
                   <Text strong>Courses:</Text>
                   <ul className="list-disc list-inside">
@@ -276,21 +203,17 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* AWARDS SECTION */}
       {awards && awards.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Awards
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Awards</Title>
+          <Divider />
           {awards.map((award, i) => (
             <div key={i} className="mb-6">
-              <Text strong>{award?.title}</Text>
+              <Text strong>{award.title}</Text>
               <br />
               <Text type="secondary">
-                {award?.awarder}, {dayjs(award?.date).format("MMM YYYY")}
+                {award.awarder}, {award.date}
               </Text>
-              {award?.summary && <p className="mt-2">{award.summary}</p>}
+              {award.summary && <p className="mt-2">{award.summary}</p>}
             </div>
           ))}
         </section>
@@ -298,24 +221,17 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* CERTIFICATES SECTION */}
       {certificates && certificates.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Certificates
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Certificates</Title>
+          <Divider />
           {certificates.map((cert, i) => (
             <div key={i} className="mb-6">
-              <Text strong>
-                {cert?.name || "Certificate name not specified"}
-              </Text>
+              <Text strong>{cert.name}</Text>
               <br />
               <Text type="secondary">
-                {cert?.issuer || "Issuer not specified"},{" "}
-                {cert?.date || "Date not specified"}
+                {cert.issuer}, {cert.date}
               </Text>
-              {cert?.url && (
+              {cert.url && (
                 <>
                   <br />
                   <a
@@ -335,24 +251,17 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* PUBLICATIONS SECTION */}
       {publications && publications.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Publications
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Publications</Title>
+          <Divider />
           {publications.map((pub, i) => (
             <div key={i} className="mb-6">
-              <Text strong>
-                {pub?.name || "Publication name not specified"}
-              </Text>
+              <Text strong>{pub.name}</Text>
               <br />
               <Text type="secondary">
-                {pub?.publisher || "Publisher not specified"},{" "}
-                {pub?.releaseDate || "Release date not specified"}
+                {pub.publisher}, {pub.releaseDate}
               </Text>
-              {pub?.url && (
+              {pub.url && (
                 <>
                   <br />
                   <a
@@ -365,7 +274,7 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
                   </a>
                 </>
               )}
-              {pub?.summary && <p className="mt-2">{pub.summary}</p>}
+              {pub.summary && <p className="mt-2">{pub.summary}</p>}
             </div>
           ))}
         </section>
@@ -373,24 +282,17 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* PROJECTS SECTION */}
       {projects && projects.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Projects
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Projects</Title>
+          <Divider />
           {projects.map((project, i) => (
             <div key={i} className="mb-6">
-              <Text strong>
-                {project?.name || "Project name not specified"}
-              </Text>
+              <Text strong>{project.name}</Text>
               <br />
               <Text type="secondary">
-                {project?.startDate || "Start date not specified"} -{" "}
-                {project?.endDate || "Ongoing"}
+                {project.startDate} - {project.endDate || "Ongoing"}
               </Text>
-              {project?.url && (
+              {project.url && (
                 <>
                   <br />
                   <a
@@ -403,10 +305,10 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
                   </a>
                 </>
               )}
-              {project?.description && (
+              {project.description && (
                 <p className="mt-2">{project.description}</p>
               )}
-              {project?.highlights && project.highlights.length > 0 && (
+              {project.highlights && project.highlights.length > 0 && (
                 <ul className="list-disc list-inside mt-2">
                   {project.highlights.map((hl, idx) => (
                     <li key={idx}>{hl}</li>
@@ -420,25 +322,15 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* SKILLS SECTION */}
       {skills && skills.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Skills
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
-          <div className="flex flex-wrap gap-2 mt-4 font-dm_sans">
+        <section className="mt-10">
+          <Title level={3}>Skills</Title>
+          <Divider />
+          <div className="flex flex-wrap gap-2 mt-4">
             {skills.map((skill) => (
-              <div
-                key={skill?.name || "unknown"}
-                className="bg-gray-100 px-3 py-1 rounded"
-              >
-                {skill?.name || "Skill name not specified"} (
-                {skill?.level || "Level not specified"})<br />
-                {skill?.keywords && skill.keywords.length > 0 && (
-                  <Text className="!font-dm_sans" type="secondary">
-                    {skill.keywords.join(", ")}
-                  </Text>
+              <div key={skill.name} className="bg-gray-100 px-3 py-1 rounded">
+                {skill.name} ({skill.level})<br />
+                {skill.keywords && skill.keywords.length > 0 && (
+                  <Text type="secondary">{skill.keywords.join(", ")}</Text>
                 )}
               </div>
             ))}
@@ -448,39 +340,30 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
 
       {/* LANGUAGES SECTION */}
       {languages && languages.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Languages
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+        <section className="mt-10">
+          <Title level={3}>Languages</Title>
+          <Divider />
           <ul className="list-none pl-0">
             {languages.map((lang, i) => (
               <li key={i} className="mb-2">
-                <Text strong>{lang?.language || "Language not specified"}</Text>{" "}
-                - <Text>{lang?.fluency || "Fluency not specified"}</Text>
+                <Text strong>{lang.language}</Text> -{" "}
+                <Text>{lang.fluency}</Text>
               </li>
             ))}
           </ul>
         </section>
       )}
+
       {/* INTERESTS SECTION */}
       {interests && interests.length > 0 && (
         <section className="mt-10">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              Interests
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
+          <Title level={3}>Interests</Title>
+          <Divider />
           <ul className="list-disc list-inside">
             {interests.map((interest, i) => (
               <li key={i}>
-                <Text strong>
-                  {interest?.name || "Interest name not specified"}
-                </Text>
-                {interest?.keywords && interest.keywords.length > 0 && (
+                <Text strong>{interest.name}</Text>
+                {interest.keywords && interest.keywords.length > 0 && (
                   <span>: {interest.keywords.join(", ")}</span>
                 )}
               </li>
@@ -488,27 +371,19 @@ const ModernTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           </ul>
         </section>
       )}
+
       {/* REFERENCES SECTION */}
       {references && references.length > 0 && (
-        <section className="mt-8">
-          <div>
-            <h3 className="text-2xl text-black font-medium font-dm_sans">
-              References
-            </h3>
-            <div className="h-0.5 w-full border-b-2 border-b-black" />
-          </div>
-          <div className="mt-2 5 grid grid-cols-2 md:grid-cols-3">
-            {references.map((ref, i) => (
-              <div key={i} className="mb-6 flex flex-col">
-                <Text className="!font-dm_sans" strong>
-                  {ref?.name || "Name not specified"}:
-                </Text>
-                <Text className="!font-dm_sans">
-                  {ref?.reference || "Reference not specified"}
-                </Text>
-              </div>
-            ))}
-          </div>
+        <section className="mt-10">
+          <Title level={3}>References</Title>
+          <Divider />
+          {references.map((ref, i) => (
+            <div key={i} className="mb-6">
+              <Text strong>{ref.name}:</Text>
+              <br />
+              <Text>{ref.reference}</Text>
+            </div>
+          ))}
         </section>
       )}
     </div>
