+++ src/components/dashboard/career/templates/StandardTemplate.tsx
@@ -110,6 +110,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           </div>
         )}
       </header>
+
       {/* SUMMARY */}
       {data.basics?.summary && (
         <section className="mb-8">
@@ -128,6 +129,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           <Text className="text-base !font-dm_sans">{data.basics.summary}</Text>
         </section>
       )}
+
       {/* WORK EXPERIENCE */}
       {data.work && data.work.length > 0 && (
         <section className="mb-8">
@@ -174,6 +176,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* EDUCATION */}
       {data.education && data.education.length > 0 && (
         <section className="mb-8">
@@ -222,6 +225,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* SKILLS */}
       {data.skills && data.skills.length > 0 && (
         <section className="mb-8">
@@ -247,6 +251,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           </div>
         </section>
       )}
+
       {/* VOLUNTEER */}
       {data.volunteer && data.volunteer.length > 0 && (
         <section className="mb-8">
@@ -289,6 +294,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* PROJECTS */}
       {data.projects && data.projects.length > 0 && (
         <section className="mb-8">
@@ -338,6 +344,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* AWARDS */}
       {data.awards && data.awards.length > 0 && (
         <section className="mb-8">
@@ -366,6 +373,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* CERTIFICATES */}
       {data.certificates && data.certificates.length > 0 && (
         <section className="mb-8">
@@ -401,6 +409,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* PUBLICATIONS */}
       {data.publications && data.publications.length > 0 && (
         <section className="mb-8">
@@ -441,6 +450,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           ))}
         </section>
       )}
+
       {/* LANGUAGES */}
       {data.languages && data.languages.length > 0 && (
         <section className="mb-8">
@@ -464,6 +474,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           </ul>
         </section>
       )}
+
       {/* INTERESTS */}
       {data.interests && data.interests.length > 0 && (
         <section className="mb-8">
@@ -489,6 +500,7 @@ const StandardTemplate: React.FC<{ data: ResumeType }> = ({ data }) => {
           </ul>
         </section>
       )}
+
       {/* REFERENCES */}
       {data.references && data.references.length > 0 && (
         <section>
